main: Make Claim* calls work before sensor is available
It's tedious for clients to know whether they can call Release* when
sensors disappear, or making sure that the device is claimed before
releasing it when they go in power save mode.

You can now claim the sensor whether it's available or not, and it will
automatically start sending out the updates when it appears.

Closes #23