data: Fix compass property name
In commit 924dd58, we changed from matching the device names to a more
generic property match. But we matched on the "iio-buffer-compass"
property instead of the one actually set in the udev rules
("iio-compass").

Change the udev rules property name to "iio-buffer-compass" so that
rules and code match.

This fixes iio-sensor-proxy startup on HP Pavilion x360.

Spotted by Ted Ying <yingted@gmail.com>

Closes: #128